+++ extras/pre-install.sh
@@ -1,6 +1,5 @@
 #!/bin/bash
-echo --------------------------------------
-echo Stopping and disabling services and daemons
+echo ---Stopping and disabling services and daemons---
 #stop services (first arancino then redis)
 systemctl stop arancino
 #systemctl stop redis-volatile
@@ -10,13 +9,12 @@ systemctl stop arancino
 systemctl disable arancino
 #systemctl disable redis-volatile
 #systemctl disable redis-persistent
-echo --------------------------------------
+echo -------------------------------------------------
 
-echo --------------------------------------
-echo Unsetting env vars
+echo ---------------Unsetting env vars----------------
 # unset envinronment vars
 unset ARANCINO
 unset ARANCINOCONF
 unset ARANCINOLOG
 unset ARANCINOENV
-echo --------------------------------------
+echo -------------------------------------------------