+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 #### v 2.4.0 2021.MM.DD
 * Introduced new Arancino Transmitter [#gn8b8f](https://app.clickup.com/t/gn8b8f)
+* Changed timestamp sent in response of `START` command. [#j1b6yc](https://app.clickup.com/t/j1b6yc)
+* Introduced Redis Time Series and the `STORE` and `STORETAGS` commands. [#acuhwp](https://app.clickup.com/t/acuhwp)
 
 #### v 2.3.0 2021.MM.DD
 * Fixed arancino config files name, now binded to ARANCINOENV var. [#g97519](https://app.clickup.com/t/g97519)