+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Updated `FLUSH` command; persistent datastore will no longer be flushed, only the volatile one: [kd64rw](https://app.clickup.com/t/kd64rw)
 * Change the Args of the `START` command: [k1518r](https://app.clickup.com/t/k1518r)
 * Introduced a new data structure used for generic attributes and a new metadata for the MCU Family: [k1518r](https://app.clickup.com/t/k1518r)
+* Introduced a new command `INTEROCEP` used to send health metrics periodically and stored in the time series: [k16dbz](https://app.clickup.com/t/k16dbz)
 
 #### v 2.3.0 2021.MM.DD
 * Fixed arancino config files name, now binded to ARANCINOENV var. [#g97519](https://app.clickup.com/t/g97519)