+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-#### v 2.4.0 2021.MM.DD
+#### v 2.4.0 2021.10.11
 * Introduced Arancino Transmitter [#gn8b8f](https://app.clickup.com/t/gn8b8f)
 * Changed timestamp format sent in response of `START` command. [#j1b6yc](https://app.clickup.com/t/j1b6yc)
 * Introduced Redis Time Series with the `STORE`, `MSTORE` and `STORETAGS` commands. [#acuhwp](https://app.clickup.com/t/k16ge0), [#acuhwp](https://app.clickup.com/t/k16ge0)
@@ -8,11 +8,11 @@
 * Introduced the Reserved Key `___BLNK_ID___` used to start the identification process on the microcontroller: blink for 5 seconds. [#jtaq9w](https://app.clickup.com/t/jtaq9w)
 * Introduced a new API endpoint to indentify a Port (if the port implements the identify process) [#jtaq9w](https://app.clickup.com/t/jtaq9w)
 * Fixed a double execution of _Arancino Commands_ [#jtb6gm](https://app.clickup.com/t/jtb6gm)
-* Introduced `MSETPERS` and `HSETPER` to store persistently even with the `mset` and `hset` api: [j594nt](https://app.clickup.com/t/j594nt)
 * Updated `FLUSH` command; persistent datastore will no longer be flushed, only the volatile one: [kd64rw](https://app.clickup.com/t/kd64rw)
 * Change the Args of the `START` command: [k1518r](https://app.clickup.com/t/k1518r)
 * Introduced a new data structure used for generic attributes and a new metadata for the MCU Family: [k1518r](https://app.clickup.com/t/k1518r)
 * Introduced a new command `INTEROCEP` used to send health metrics periodically and stored in the time series: [k16dbz](https://app.clickup.com/t/k16dbz)
+* Fix: API Set Config misses authentication. [jz5mef](https://app.clickup.com/t/jz5mef)
 
 #### v 2.3.0 2021.MM.DD
 * Fixed arancino config files name, now binded to ARANCINOENV var. [#g97519](https://app.clickup.com/t/g97519)